Initial Commit
Basic python game with a boss that randomly moves left and right on top.  The boss fires projectiles that can either hit blocks on playing field, hit the player or miss and be destroyed.  The objective of the game is to kill the boss with arrows and dodge incoming attacks.  The game runs in a while loop since I found it to be the only way to replay.  If the boss or player is killed, the player is brought to the main menu.